Move `Iterate` into `solver`
